- Update based on github comments - Fixed the command arguments, replace "_" with "-" - Removed old_xmltemplate.py - Update .coveragerc - Add more entries in YEARS_FOR_CYCLE under citations/__init__.py - Clean up setup.md - Clean up queue_manager/config.py - Add a step to clean up staging directory before pipeline process is completed

+++ HST/organize_files.py
@@ -36,4 +36,21 @@ def organize_files_from_staging_to_bundles(proposal_id, logger):
                 os.makedirs(bundles_prod_dir, exist_ok=True)
                 logger.info(f'Move {dir} from staging to bundles directory')
                 shutil.copytree(staging_prod_dir, bundles_prod_dir, dirs_exist_ok=True)
+
+def clean_up_staging_dir(proposal_id, logger):
+    """Remove organized directories from the staging folder (they are copied to the
+    bundle directory). This step should be run at the end of the pipeline process.
+
+    Inputs:
+        proposal_id    a proposal id.
+        logger         pdslogger to use; None for default EasyLogger.
+    """
+    logger = logger or pdslogger.EasyLogger()
+    logger.info(f'Clean up staging directory for proposal id: {proposal_id}')
+    staging_dir = get_program_dir_path(proposal_id, None, root_dir='staging')
+    for dir in os.listdir(staging_dir):
+        for col_prefix in COL_NAME_PREFIX:
+            if dir.startswith(col_prefix):
+                staging_prod_dir = os.path.join(staging_dir, dir)
+                logger.info(f'Remove {dir} from staging directory')
                 shutil.rmtree(staging_prod_dir)
